need three list to properly reference a hardware
we need a list for :

* devices_name: as obtained from EnumerateUSB ofr instance
* dlls: possible dlls for the PI hardware, see the config of the plugin
* device: both a device_name and a corresponding dll

I have to do that last bit because different dlls can enumerate devices names but not run it. For instance I have an old E-625 controller drived by the E816 dll. This device is found both with the GCS2 dll and the E816 dll, but only the last one will be able to drive the motor...